move tremor utils to the correct folder
